+++ components/nav.tsx
@@ -2,9 +2,10 @@ import styled from "styled-components";
 import Image from "next/image";
 import { useEffect, useState } from "react";
 import { SearchBtn } from "../styles/globalStyle";
+import Link from "next/link";
 const nav = () => {
   const [scrollState, setScrollState] = useState<number>(0);
-  const [navVisibility, setNavVisibility] = useState<number>(1)
+  const [navVisibility, setNavVisibility] = useState<number>(1);
 
   // console.log("랜더링")
   // console.log(scrollState)
@@ -15,12 +16,18 @@ const nav = () => {
           <Image src={"/icon/brandLogo.png"} width={150} height={110}></Image>
         </Logo>
         <UlContainer>
-          <LiItem className="hover">반려동물 이야기</LiItem>
+          <LiItem className="hover">
+            <Link href="/talk">반려동물 이야기</Link>
+          </LiItem>
           <LiItem className="hover">예약하기</LiItem>
           <LiItem className="hover">
-            <Image src={"/backgroundImage/addMarket.png"} width={20} height={20}></Image>
+            <Image
+              src={"/backgroundImage/addMarket.png"}
+              width={20}
+              height={20}
+            ></Image>
             매장등록
-            </LiItem>
+          </LiItem>
           <SearchBox>
             <SearchBoxInput placeholder="어떤 서비스를 찾고 계신가요?"></SearchBoxInput>
             <SearchBtn className="hover">검색</SearchBtn>
@@ -53,7 +60,7 @@ const nav = () => {
 export default nav;
 
 const MainNavContainer = styled.div.attrs(() => {})`
-  display:flex;
+  display: flex;
   width: 100%;
   height: 120px;
   color: white;
@@ -64,7 +71,7 @@ const MainNavContainer = styled.div.attrs(() => {})`
     width: 100%;
     height: 100vh;
     content: "";
-    background-color: rgb(181, 214, 146,0.8);
+    background-color: rgb(181, 214, 146, 0.8);
     /* filter: brightness(70%); */
     position: absolute;
     top: 0;
@@ -79,7 +86,6 @@ const MainNavContainer = styled.div.attrs(() => {})`
   z-index: 1;
   overflow: hidden;
   padding-left: 100px;
-
 
   @media screen and (max-width: 768px) {
     height: 70px;
@@ -97,7 +103,7 @@ const Logo = styled.div`
   margin-top: 10px;
   margin-left: -5px;
   font-weight: bold;
-  color:black;
+  color: black;
 
   @media screen and (max-width: 1024px) {
     margin-top: 6px;
@@ -123,19 +129,18 @@ const UlContainer = styled.ul`
 const LiItem = styled.li`
   &:hover {
     border-top: 2px solid orange;
-    padding-top:2px;
+    padding-top: 2px;
   }
   margin-right: 18px;
-  &:nth-child(1){
+  &:nth-child(1) {
     color: black;
-
   }
   &:nth-child(2) {
     color: black;
   }
 
   &:nth-child(3) {
-    color:rgb(102,26,0,1);
+    color: rgb(102, 26, 0, 1);
   }
 `;
 
@@ -156,13 +161,12 @@ export const SearchBoxInput = styled.input`
   height: 28px;
   border: none;
   border-radius: 10px;
-  padding-left:6px;
+  padding-left: 6px;
 
   &:focus {
     outline: 2px solid orange;
   }
 `;
-
 
 const MyPageToolBox = styled.div`
   display: flex;
@@ -172,7 +176,7 @@ const MyPageToolBox = styled.div`
   margin-top: -5px;
   @media screen and (max-width: 1024px) {
     margin-left: 0;
-    margin-right:77px;
+    margin-right: 77px;
 
     margin-top: 34px;
     width: 33px;
@@ -195,14 +199,13 @@ const MobileNavContainer = styled.div.attrs(() => {})`
   overflow: hidden;
   padding-left: 20px;
   padding-right: 20px;
-  background-color:rgb(181, 214, 146,0.9);
+  background-color: rgb(181, 214, 146, 0.9);
 
   @media screen and (max-width: 1024px) {
     display: flex;
     width: 100%;
     justify-content: space-between;
   }
- 
 `;
 
 const MobileMenu = styled.div`
